+++ src/actions/index.js
@@ -1,32 +1,11 @@
-import { ADD_MESSAGE, INITIAL_DATA, ADD_USER, USERS_LIST, MESSAGE_RECEIVED } from '../constants/actionsTypes';
+import { initialLoad } from './common';
+import { addUser, populateUsersList } from './users';
+import { addMessage, messageReceived } from './messages';
 
-export const initialLoad = ({ users, messages }) => ({
-  type: INITIAL_DATA,
-  users,
-  messages
-});
-
-export const addMessage = (message, uuid, uuidAuthor, author) => ({
-  type: ADD_MESSAGE,
-  uuid,
-  message,
-  uuidAuthor,
-  author
-});
-
-export const addUser = name => ({
-  type: ADD_USER,
-  name
-});
-
-export const messageReceived = (message, uuid, author) => ({
-  type: MESSAGE_RECEIVED,
-  uuid,
-  message,
-  author
-});
-
-export const populateUsersList = users => ({
-  type: USERS_LIST,
-  users
-});
+export {
+  initialLoad,
+  addUser,
+  populateUsersList,
+  addMessage,
+  messageReceived
+};